// src/services/types.ts
export interface NavLink {
  path: string
  label: string
  icon: React.ReactNode
}

export interface SidebarProps {
  open: boolean
  setOpen: (open: boolean) => void
}